muh safety

+++ ci/bootstrap.sh
@@ -53,6 +53,17 @@ CLONE_URI="git@github.com:spatialos/gdk-for-unity.git"
 TARGET_DIRECTORY="$(realpath $(pwd)/../gdk-for-unity)"
 PINNED_VERSION=$(cat ./gdk.pinned)
 
+if [[ -z ${BUILDKITE:-} ]]; then
+    echo "Warning: About to delete ${TARGET_DIRECTORY}. Please confirm. (Default is Cancel)"
+    read -p "Y/N > " -r
+    echo    # (optional) move to a new line
+    if [[ $REPLY =~ ^[Yy]$ ]]; then
+        echo "Deleting..."
+    else
+        exit 1
+    fi
+fi
+
 rm -rf "${TARGET_DIRECTORY}"
 
 mkdir "${TARGET_DIRECTORY}"
